requirements.txt
diffusers==0.36.0
torch==2.7.0
transformers==5.1.0
pillow
numpy
xformers==0.0.30
spaces
accelerate
imageio
peft==0.10.0
pyyaml
opencv-python
awscli
boto3
tqdm